+++ sgp4-predict-py/python/sgp4_predict/__init__.py
@@ -18,6 +18,7 @@ from sgp4_predict._sgp4_predict import (
     Illumination,
     IlluminationIter,
     IlluminationState,
+    Interval,
     LatLon,
     Observation,
     ObservationIter,
@@ -49,25 +50,6 @@ class IntervalRange(Protocol):
 
     @property
     def end(self) -> datetime: ...
-
-
-class Interval:
-    """Concrete datetime interval satisfying IntervalRange."""
-
-    def __init__(self, start: datetime, end: datetime) -> None:
-        self._start = start
-        self._end = end
-
-    @property
-    def start(self) -> datetime:
-        return self._start
-
-    @property
-    def end(self) -> datetime:
-        return self._end
-
-    def __repr__(self) -> str:
-        return f"Interval(start={self._start}, end={self._end})"
 
 
 __all__ = [